use different version to get info_update2, explanation in long desc
The guide previously did not indicate that cake is not a legitimately obtainable block in 0.6.0, nor that downgrading is required to convert cake to info_update2, which could cause readers to incorrectly conclude that cake converts to info_update2 upon updating from 0.6.0 to 0.7.1. Instead, we use 0.8.0 -> 0.8.0_test1 to obtain the block as that's a downgrade already built into the guide. Since cake does not convert to info_update2 using this downgrade, we swap it out for beetroot crops.