multiple changes, explanation in long desc
The glitch allowing for no nbt enchanted books was patched in 0.12.1-b6, so that has been moved to an earlier version. Build 3 was chosen as it is the build where the player is the least likely to encounter any annoying bugs.

Sponges are not legitimately obtainable in 0.12.1, so a 0.16.0-b4 -> 0.14.3 downgrade has been added to the version path to allow the player to properly obtain dry sponges in water. Why not 0.16.0-b4 -> 0.15.9? Because the game crashes on startup if you try doing that.